[step2] feat : 로또복권 생성기능을 LottoTicketGenerator로 위임
